+++ fast-exp/src/fast_exp.hpp
@@ -1,15 +1,16 @@
 #ifndef FAST_EXP_HPP
 #define FAST_EXP_HPP
 
+#include <cstdint>
+
 /**
  * Computes a^p mod n
  *
  * @param a
  * @param p
  * @param n
- * @return
+ * @return a^p mod n
  */
-int fastExp(int a, int p, int n);
+uint64_t fastExp(uint64_t a, uint64_t p, uint64_t n);
 
 #endif
-
